refactor(web): isolate day calendar scroll command
